finished color coding cards
