+++ index.html
@@ -12,7 +12,7 @@
     <div id="parkInfo"></div>
     <div id="parkAreas"></div>
     <div id="attractionType"></div>
-    <div id="attractionList" class="row card-group"></div>
+    <div id="attractionList" class="row "></div>
     <script src="./dist/app.js"></script>
 </body>
 </html>
